+++ Libretti/include/Effects.h
@@ -4,6 +4,8 @@
 *
 * Effects
 *
+* Contains structs with control parameters for effects.
+*
 * Copyright (c) 2017-2018 Jean-Louis Haywood. All rights reserved.
 * License: https://github.com/Jean-LouisH/Libretti/blob/master/LICENSE
 */
@@ -13,26 +15,26 @@
 
 typedef struct
 {
-	double preDelay_ms;
-	double roomSize_pct;
-	double damping_pct;
+	float preDelay_ms;
+	float roomSize_pct;
+	float damping_pct;
 }ReverbControl;
 
 typedef struct
 {
-	double delay_ms;
-	double decay_pct;
+	float delay_ms;
+	float decay_pct;
 }EchoControl;
 
 typedef struct
 {
-	double level_dB;
+	float level_dB;
 }EQControl;
 
 typedef struct
 {
-	double extent_cents;
-	double rate_per_s;
+	float extent_cents;
+	float rate_per_s;
 }VibratoControl;
 
 typedef struct
@@ -42,7 +44,7 @@ typedef struct
 	EQControl eq[10]; //31Hz, 62Hz, 125Hz, 250Hz, 500Hz, 1kHz, 2kHz, 4kHz, 8kHz, 16kHz
 	VibratoControl vibrato;
 	uint32_t crossfading_ms;
-	double pitchBlend_pct;
+	float pitchBlend_pct;
 }lb_Effects;
 
 #endif /*Effects_h*/
